Mise à jour des contrôleurs et modèles liés à l'authentification (login, inscription, vérification d'email, etc.)

+++ app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -5,71 +5,46 @@ namespace App\Http\Controllers\Auth;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Auth\LoginRequest;
 use App\Models\User;
+use Hash;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
-use Illuminate\Support\Facades\Auth;
-use Validator;
+use App\Http\Resources\Auth\AuthLoginResource;
 
 class AuthenticatedSessionController extends Controller
 {
     /**
      * Handle an incoming authentication request.
      */
-    public function store(Request $request): JsonResponse
-    {
-        // Validation des données
-        $validated = Validator::make($request->all(), [
-            'email' => 'required|string|email:rfc',
-            'password' => 'required|string|min:6'
-        ]);
+    public function store(LoginRequest $request): AuthLoginResource | JsonResponse
+        {
+            $user = User::where('email', '=', $request->validated('email'))->first();
 
-        // Captures des erreurs 
-        if ($validated->fails()) {
-            return response()->json(['error' => $validated->errors()], 403);
-
-        }
-
-        $credentials = ['email' => $request->email, 'password' => $request->password];
-
-        try {
-
-            if (!auth()->attempt($credentials)) {
-                return response()->json(['error' => 'email or password incorrect '], 400);
+            if (! ($user instanceof User) || !Hash::check($request->validated('password'), $user->password)) {
+                return response()->json(['message' => 'Unauthorized'], 401);
             }
 
-            $user = User::where('email', $request->email)->firstOrFail();
-            
-            $token = $user->createToken('token')->plainTextToken;
-            $user['token'] = $token;
 
-            return response()->json([
-                'data' => $user,
-            ], 201);
+            $token = $user->createToken($user->email)->plainTextToken;
 
-        } catch (\Exception $exception) {
-            return response()->json([
-                'error' => [
-                    $exception->getMessage()
-                ]
-            ], 500);
+            $user->token = $token;
+
+            return new AuthLoginResource($user);
+
         }
-
-    }
 
     /**
      * Destroy an authenticated session.
      */
     public function destroy(Request $request): JsonResponse
     {
-        try {
-            //currentAccessToken pour supprimer que le token de la session en cours
-            $request->user()->currentAccessToken()->delete();
-            return response()->json(['message' => 'deconnexion reussie']);
+        $user  = $request->user();
 
-        } catch (\Exception $e) {
-            return response()->json(['error' => $e->getMessage()]);
+        if (!($user instanceof User)) {
+            return response()->json(['message' => 'Unauthorized'], 401);
         }
 
+        $request->user()->currentAccessToken()->delete();
+
+        return response()->json(['logout' => true]);
     }
 }
